+++ R/inflate-utils.R
@@ -366,6 +366,20 @@ group_code <- function(df, group_col, code_col) {
     do.call("rbind", .)
 }
 
+#' Retrieve name of the package if there is a DESCRIPTION
+#' @noRd
+get_pkg_name <- function(pkg){
+  desc <- file.path(pkg, "DESCRIPTION")
+  if (file.exists(desc)) {
+    pkgname <- read.dcf(desc)[colnames(read.dcf(desc)) == "Package"]
+  } else{
+    pkgname <- basename(pkg)
+
+  }
+  pkgname
+}
+
+
 #' Create vignette header
 #' @param pkg Path to package
 #' @param vignette_name Name of the resulting vignette
@@ -374,7 +388,7 @@ group_code <- function(df, group_col, code_col) {
 #' @importFrom utils getFromNamespace
 #' @noRd
 create_vignette_head <- function(pkg, vignette_name, yaml_options = NULL) {
-  pkgname <- basename(pkg)
+  pkgname <- get_pkg_name(pkg = pkg)
 
   # Get all yaml options except Title, output, editor_options
   yaml_options <- yaml_options[