+++ tests/testthat/test-inflate_utils.R
@@ -24,7 +24,7 @@ test_that("group_code groups columns", {
 dummypackage <- tempfile("isrproj.pkg")
 dir.create(dummypackage)
 
-
+# {fusen} steps
 fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
 dev_file <- suppressMessages(add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE))
 flat_file <- dev_file[grepl("flat_", dev_file)]
@@ -40,7 +40,7 @@ usethis::with_project(dummypackage, {
   test_that("is_pkg_proj works when rproj no pkg", {
     expect_equal(is_pkg_proj(dummypackage), FALSE)
   })
-
+  # create_vignette
   usethis::with_project(dummypackage, {
     inflate(pkg = dummypackage, flat_file = flat_file,
             vignette_name = "Get started", check = FALSE,
@@ -138,3 +138,39 @@ test_that("create_vignette_head works", {
   })
 })
 unlink(dummypackage, recursive = TRUE)
+
+
+# get_pkg_name ----
+dummypackage <- tempfile("checkpk.gname2")
+dir.create(dummypackage)
+# rstudioapi::filesPaneNavigate(dummypackage)
+# {fusen} steps
+fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package", Package = "COUCOU2"))
+dev_file <- suppressMessages(add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE))
+flat_file <- dev_file[grepl("flat_", dev_file)]
+
+test_that("get_pkg_name works", {
+  flat_file_lines <- readLines(flat_file)
+  #the package name must be used and not the project name
+  expect_true(any(grepl("COUCOU2", flat_file_lines, fixed = TRUE)))
+  expect_false(any(grepl("gname2", flat_file_lines)))
+})
+
+unlink(dummypackage, recursive = TRUE)
+
+
+dummypackage <- tempfile("checkpk.gname")
+dir.create(dummypackage)
+# {fusen} steps
+fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package",Package="COUCOU"))
+dev_file <- suppressMessages(add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE))
+flat_file <- dev_file[grepl("flat_", dev_file)]
+
+test_that("get_pkg_name inflates", {
+  usethis::with_project(dummypackage, {
+    inflate(pkg = dummypackage, flat_file = flat_file,
+            vignette_name = "Get started", check = FALSE,
+            open_vignette = FALSE)
+  })
+})
+